refactor(app): move RunTimer to molecules
