+++ app/src/organisms/Devices/ProtocolRun/ProtocolRunHeader/RunHeaderContent/RunHeaderSectionUpper.tsx
@@ -11,7 +11,7 @@ import {
 
 import { LabeledValue } from './LabeledValue'
 import { DisplayRunStatus } from '../DisplayRunStatus'
-import { RunTimer } from '../../RunTimer'
+import { RunTimer } from '/app/molecules/RunTimer'
 import { ActionButton } from './ActionButton'
 import { useRunCreatedAtTimestamp } from '../../../hooks'
 import { useRunTimestamps } from '/app/resources/runs'
